HZN-1180: Minor changes
