+++ features/telemetry/listeners/flow/pom.xml
@@ -41,10 +41,6 @@
       <artifactId>commons-csv</artifactId>
     </dependency>
     <dependency>
-      <groupId>com.google.protobuf</groupId>
-      <artifactId>protobuf-java</artifactId>
-    </dependency>
-    <dependency>
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
       <scope>test</scope>
